requirements.txt
django-auth-ldap==2.2.0
PyJWT==2.0.1
python-ldap==3.3.1
Django~=3.1.5
requests~=2.22.0
gunicorn~=20.0.4